// enumeratewebapp/src/model/UserInfo.ts
export interface UserInfo {
    userId: number;
    userName: string;
    name: string;
    pid: string;
    email: string;
    phone: string;
    positionId: number;
    positionName: string;
    roleId: number;
    roleName: string;
    reg?: string;
    regn?: string;
    cwtCode?: string;
    cwtName?:string;
    ampCode?: string;
    ampName?:string
    tamCode?: string;
    tamName?:string
    typeCode?: string;
    munCode?: string;
    munName?:string
    vilCode?: string; 
    vilName?:string;
    vilt?:string;
    aH_CODE?: string;
    areA_CODE?: string;
}