// enumeratewebapp/src/model/TriangulationInfo.ts
export interface TriangulationInfo {
  TRIANGULATION_KEY: number;
  AREA_CODE: string;
  REG: string;
  REGN: string;
  CWT: string;
  CWTN: string;
  AMP: string;
  AMPN: string;
  TAM: string;
  TAMN: string;
  TYPE: string;
  MUN: string;
  MUNN: string;
  VIL: string;
  VILN: string;
  VILT: string;
  TotalAH_NO: number;
  TotalPlant: number;
  TotalShieling: number;
  TotalAquafarming: number;
  TotalSeaSalt: number;
  TotalA07_AVG: number;
  TotalA07_MAX: number;
  TotalA07_MIN: number;
  TotalINCOME_AVG: number;
  TotalMEMBER_AVG: number;
  TotalDEBT_AVG: number;
  ENU_TotalAH_NO: string;
  ACA_TotalAH_NO: string;
  SUB_TotalAH_NO: string;
  ENU_TotalActivity: string;
  SUB_TotalActivity: string;
  ACA_TotalActivity: string;
  ENU_TotalA07_AVG: string;
  ACA_TotalA07_AVG: string;
  SUB_TotalA07_AVG: string;
  ENU_TotalA07_MAX: string;
  ACA_TotalA07_MAX: string;
  SUB_TotalA07_MAX: string;
  ENU_TotalA07_MIN: string;
  ACA_TotalA07_MIN: string;
  SUB_TotalA07_MIN: string;
  ENU_TotalINCOME_AVG: string;
  ACA_TotalINCOME_AVG: string;
  SUB_TotalINCOME_AVG: string;
  ENU_TotalDEBT_AVG: string;
  ACA_TotalDEBT_AVG: string;
  SUB_TotalDEBT_AVG: string;
  ENU_TotalMEMBER_AVG: string;
  ACA_TotalMEMBER_AVG: string;
  SUB_TotalMEMBER_AVG: string;
  ENU_NAME: string;
  ACA_NAME: string;
  SUB_NAME: string;
  ENU_REMARK: string;
  ACA_REMARK: string;
  SUB_REMARK: string;
  ENU_APPROVE_DATE: Date | null;
  ACA_APPROVE_DATE: Date | null;
  SUB_APPROVE_DATE: Date | null;
  Lock: string;
}